+++ topkio-service/src/handlers.rs
@@ -1,0 +1,2 @@
+mod chat_completion;
+pub use chat_completion::handle_chat_completion;
